Make tenant attribute immutable only if the current tenant is set.
It's need to make tenant attribute immutable only if the current tenant
is set. Sometimes a system area doesn't have a tenant set, needs to edit
that attribute. An example:

When you are in a administrative area and a tenant is not set. You need be able
to edit the tenant attribute.