Setting tenant_id to a string might be okay
In Ruby, String#to_i returns an integer if the string begins with an
integer.

    2.2.2 :001 > "1-microsoft-way".to_i
     => 1

Rails takes advantage of this behavior and does not bother converting
fields such as `tenant_id` from String to Integer types. We should allow
these values to be passed into the `"#{fkey}="` methods as well.

P.S. There was no `integer` in scope on line 102. Tests are important
:D