Remove getter override method (#7)
* Remove getter override method

There are negligible performance gains in caching the tenant object to
justify the complexity.

Removing the cache lets ActiveRecord return the fresh object and it uses
internal caching when accessing it from the same object.

* Remove Account class that overrides the ActiveRecord one

This class already comes from `spec/active_record_models`. By overriding
it, when all files are loaded, this one ends up creating some kind of
conflict with the ActiveRecord one, and not pushing the values down to
the database as it is required to other tests to work (foreign key and
primary key ones).

* Let CustomPrimaryKeyTask class have its own table

No real reason to point to another table just because they share one of
the same attribute: `name`.

It was more confusing to debug because `projects` table has `account_id`
attribute, which is not used and not what CustomPrimaryKeyTask
represents. It's there because ActiveRecord loads the attributes from
the table.

* Add expectation to test

Changing it from == to != does not fail the test.

* Honour dependency on Active Record models for all tests

There were controller tests that were creating its own Account so that
they could run in isolation.

However, due to interdependency between files, it affected other tests.

As of now interdependency still exists and I aim to eliminate soon.

* Refrain from depending on mocking and another Account implementation

Added missing `_spec` suffix to test file.

Made tests more resilient by exercising ActiveRecord. These tests were
highly coupled to the implementation. If the implementation changes, the
tests would break.